Use a custom format+parser for commands
This avoids a bunch of issues we were having with using a strict xml
parser. We don't have to worry about claude escaping inner fields
properly.

Instead we use a magic string to mark the commands. The string is
unlikely to appear in the content of programs and does not appear
in the content of this program so we can run code-buddy against
itself.